[TechDraw] Move CosmeticVertex into own file
