+++ src/Mod/TechDraw/App/CosmeticVertexPyImp.cpp
@@ -32,6 +32,7 @@
 #include <Base/VectorPy.h>
 
 #include "Cosmetic.h"
+#include "CosmeticVertex.h"
 #include "CosmeticVertexPy.h"
 #include "CosmeticVertexPy.cpp"
 #include "DrawUtil.h"
